Refactoring: Create new PasswordRequirement and UsernameRequirement
